+++ app/src/main/java/com/ost/mge/inventoryapp/SearchView.kt
@@ -1,0 +1,69 @@
+package com.ost.mge.inventoryapp
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.RectangleShape
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun SearchView(searchState: MutableState<TextFieldValue>) {
+    TextField(
+        value = searchState.value,
+        onValueChange = {value -> searchState.value = value },
+        modifier = Modifier.fillMaxWidth(),
+        singleLine = true,
+        shape = RectangleShape,
+        textStyle = TextStyle(color = Color.White, fontSize = 18.sp),
+        leadingIcon = {
+            Icon(
+                Icons.Default.Search,
+                contentDescription = "Search categories",
+                modifier = Modifier
+                    .padding(15.dp)
+                    .size(22.dp)
+            )
+        },
+        trailingIcon = {
+            if (searchState.value != TextFieldValue("")) {
+                IconButton(
+                    onClick = {
+                        searchState.value =
+                            TextFieldValue("")
+                    }
+                ) {
+                    Icon(
+                        Icons.Default.Close,
+                        contentDescription = "Clear search",
+                        modifier = Modifier
+                            .padding(15.dp)
+                            .size(24.dp)
+                    )
+                }
+            }
+        },
+//        colors = TextFieldDefaults.textFieldColors(
+//            textColor = Color.White,
+//            cursorColor = Color.White,
+//            leadingIconColor = Color.White,
+//            trailingIconColor = Color.White,
+//            backgroundColor = colorResource(id = R.color.colorPrimary),
+//            focusedIndicatorColor = Color.Transparent,
+//            unfocusedIndicatorColor = Color.Transparent,
+//            disabledIndicatorColor = Color.Transparent
+//        )
+    )
+}
